Better tests for Intervals. API change to set:
The behaviour before was that dates were reset to 0 on the second `set` call.
This has been removed in favour of the date only being reset on first set. This
allows Tempus ISO Intervals to set on top of itself, and it doesn't have to rely
on making it's own dates before setting.

The "downside" to this is that when Tempus ISO Intervals called with just an
interval string will now set on Epoch. Not really a downside though considering
they should have done that all along.