Fix Tempus.Timer to work with new interval change
This also fixes an undiscovered issue with `Tempus.Timer.on()` which was not
instantiating Tempus with multiple args. Now anything you pass to on() will get
passed to Tempus().

It also fixed another undiscovered issue whereby Tempus was passing `arguments`
to modules, when it should be passing it's de-munged `ar` variable.